network module: guard onboot configuration task by environment
